chore: move v9-specific logic into its own separate module.
